Move Profile to src/auth
